+++ src/components/views/home/applianceModal/appliance/OccurenceButtons.js
