+++ src/components/views/home/applianceModal/appliance/occurence/DurationSelector.js
